chore: apply actions (#379)
* chore: apply new actions

* chore: remove unused part

* chore: set global error log variable

* chore: change action name

* chore: apply version
