--- .github/workflows/examplePageTest.yml
@@ -1,29 +1,0 @@
-name: Node.js CI
-
-on:
-  schedule:
-    - cron: '0 22 * * *'
-
-jobs:
-  build:
-
-    runs-on: ubuntu-latest
-
-    strategy:
-      matrix:
-        node-version: [12.x]
-
-    steps:
-    - uses: actions/checkout@v2
-    - name: Use Node.js ${{ matrix.node-version }}
-      uses: actions/setup-node@v1
-      with:
-        node-version: ${{ matrix.node-version }}
-    - run: npm install
-    - run: ROWSERSTACK_ACCESS_KEY=$BROWSERSTACK_ACCESS_KEY
-    - run: BROWSERSTACK_USERNAME=$BROWSERSTACK_USERNAME
-    - run: node exampleTest.js
-      env:
-        CI: true
-        BROWSERSTACK_ACCESS_KEY: ${{ secrets.BROWSERSTACK_ACCESS_KEY }}
-        BROWSERSTACK_USERNAME: ${{ secrets.BROWSERSTACK_USERNAME }}
